+++ lib/hitobito_sjas.rb
@@ -7,7 +7,7 @@
 
 
 require 'wagons'
-# require 'your_wagon_dependencies'
+require 'hitobito_youth'
 require 'hitobito_sjas/wagon'
 
 module HitobitoSjas